+++ clang-tools-extra/clang-tidy/openmp/OpenMPTidyModule.cpp
@@ -17,6 +17,7 @@
 #include "ImplementationDefinedUnrollCheck.h"
 #include "MissingForInParallelDirectiveBeforeForLoopCheck.h"
 #include "MissingOrderedDirectiveAfterOrderedClauseCheck.h"
+#include "RecursiveTaskWithoutTaskCreationConditionCheck.h"
 #include "SpecifyScheduleCheck.h"
 #include "TaskDependenciesCheck.h"
 #include "UnprotectedSharedVariableAccessCheck.h"
@@ -49,6 +50,8 @@ public:
             "openmp-missing-for-in-parallel-directive-before-for-loop");
     CheckFactories.registerCheck<MissingOrderedDirectiveAfterOrderedClauseCheck>(
         "openmp-missing-ordered-directive-after-ordered-clause");
+    CheckFactories.registerCheck<RecursiveTaskWithoutTaskCreationConditionCheck>(
+        "openmp-recursive-task-without-task-creation-condition");
     CheckFactories.registerCheck<TaskDependenciesCheck>(
         "openmp-task-dependencies");
     CheckFactories.registerCheck<UseAtomicInsteadOfCriticalCheck>(